gitmodules: Use relative paths for the tt-metal hosted repos
This means that if you clone with https://github.com/tenstorrent-metal/tt-metal,
the submodules will also use https. Slight removal of friction for
those not using SSH for github (cloning).

Main drawback is that if you for tt-metal, you'd need to fork the
submodule repos too since they're relative. If that's considered a higher
friction point, then this is the wrong approach. I'll leave that up to
project maintainers to decide, just wanted to propose the change.

Signed-off-by: Olof Johansson <olofj@tenstorrent.com>